v0.0.6 - Fixed wrong 'index' instead of 'humandiff' in readme's USAGE section.
This was caused because this section is generated from actual binary
help output, but binary uses its calling name as program name which in
the project is 'index'(.js) and symlink generated in 'npm install -g'
installation is 'humandiff'.

I renamed entry poing file (index.js) to 'humandiff.js' to avoid this to
happen again in the future.

Also take the opportunity to update / improve advanced features
explanation.